Move app to own file; clippy
