+++ Cargo.toml
@@ -48,6 +48,6 @@ color-eyre = "0.6.3"
 crossterm = "0.28.1"
 copypasta = "0.10.1"
 regex = "1.11.1"
+insta = "1.41.1"
 
 chrono = { version = "0.4.39", features = ["serde", "clock"] }
-
